chore(refactor): refactoring code to reduce cyclical imports in order to prep for migration to rollup + vite (which have much more strict module resolutions)
