+++ .changeset/khaki-sheep-bake.md
@@ -1,0 +1,5 @@
+---
+"electron-updater": patch
+---
+
+chore(refactor): refactoring code to reduce cyclical imports in order to migrate to rollup + vite (which have much more strict module resolutions)
