+++ packages/electron-updater/src/MacUpdater.ts
@@ -5,7 +5,8 @@ import * as path from "path"
 import { createServer, IncomingMessage, Server, ServerResponse } from "http"
 import { AppAdapter } from "./AppAdapter"
 import { AppUpdater, DownloadUpdateOptions } from "./AppUpdater"
-import { ResolvedUpdateFileInfo, UpdateDownloadedEvent } from "./main"
+import { ResolvedUpdateFileInfo } from "./main"
+import { UpdateDownloadedEvent } from "./exports"
 import { findFile } from "./providers/Provider"
 import AutoUpdater = Electron.AutoUpdater
 import { execFileSync } from "child_process"
