+++ packages/electron-updater/src/main.ts
@@ -1,14 +1,10 @@
-import { CancellationToken, PackageFileInfo, ProgressInfo, UpdateFileInfo, UpdateInfo } from "builder-util-runtime"
-import { EventEmitter } from "events"
 import { existsSync, readFileSync } from "fs-extra"
 import * as path from "path"
-import { URL } from "url"
 import { AppUpdater } from "./AppUpdater"
-import { LoginCallback } from "./electronHttpExecutor"
+import { UpdateInfo } from "builder-util-runtime"
 
 export { BaseUpdater } from "./BaseUpdater"
 export { AppUpdater, NoOpLogger } from "./AppUpdater"
-export { CancellationToken, PackageFileInfo, ProgressInfo, UpdateFileInfo, UpdateInfo }
 export { Provider } from "./providers/Provider"
 export { AppImageUpdater } from "./AppImageUpdater"
 export { DebUpdater } from "./DebUpdater"
@@ -16,6 +12,8 @@ export { PacmanUpdater } from "./PacmanUpdater"
 export { RpmUpdater } from "./RpmUpdater"
 export { MacUpdater } from "./MacUpdater"
 export { NsisUpdater } from "./NsisUpdater"
+
+export * from "./exports"
 
 // autoUpdater to mimic electron bundled autoUpdater
 let _autoUpdater: any
@@ -68,83 +66,6 @@ Object.defineProperty(exports, "autoUpdater", {
     return _autoUpdater || doLoadAutoUpdater()
   },
 })
-
-export interface ResolvedUpdateFileInfo {
-  readonly url: URL
-  readonly info: UpdateFileInfo
-
-  packageInfo?: PackageFileInfo
-}
-
-export interface UpdateCheckResult {
-  readonly isUpdateAvailable: boolean
-
-  readonly updateInfo: UpdateInfo
-
-  readonly downloadPromise?: Promise<Array<string>> | null
-
-  readonly cancellationToken?: CancellationToken
-
-  /** @deprecated */
-  readonly versionInfo: UpdateInfo
-}
-
-export type UpdaterEvents = "login" | "checking-for-update" | "update-available" | "update-not-available" | "update-cancelled" | "download-progress" | "update-downloaded" | "error"
-
-export const DOWNLOAD_PROGRESS = "download-progress"
-export const UPDATE_DOWNLOADED = "update-downloaded"
-
-export type LoginHandler = (authInfo: any, callback: LoginCallback) => void
-
-export class UpdaterSignal {
-  constructor(private emitter: EventEmitter) {}
-
-  /**
-   * Emitted when an authenticating proxy is [asking for user credentials](https://github.com/electron/electron/blob/master/docs/api/client-request.md#event-login).
-   */
-  login(handler: LoginHandler): void {
-    addHandler(this.emitter, "login", handler)
-  }
-
-  progress(handler: (info: ProgressInfo) => void): void {
-    addHandler(this.emitter, DOWNLOAD_PROGRESS, handler)
-  }
-
-  updateDownloaded(handler: (info: UpdateDownloadedEvent) => void): void {
-    addHandler(this.emitter, UPDATE_DOWNLOADED, handler)
-  }
-
-  updateCancelled(handler: (info: UpdateInfo) => void): void {
-    addHandler(this.emitter, "update-cancelled", handler)
-  }
-}
-
-export interface UpdateDownloadedEvent extends UpdateInfo {
-  downloadedFile: string
-}
-
-const isLogEvent = false
-
-function addHandler(emitter: EventEmitter, event: UpdaterEvents, handler: (...args: Array<any>) => void): void {
-  if (isLogEvent) {
-    emitter.on(event, (...args: Array<any>) => {
-      console.log("%s %s", event, args)
-      handler(...args)
-    })
-  } else {
-    emitter.on(event, handler)
-  }
-}
-
-export interface Logger {
-  info(message?: any): void
-
-  warn(message?: any): void
-
-  error(message?: any): void
-
-  debug?(message: string): void
-}
 
 /**
  * return null if verify signature succeed
