+++ packages/electron-updater/src/providers/GitHubProvider.ts
@@ -2,7 +2,7 @@ import { CancellationToken, GithubOptions, githubUrl, HttpError, newError, parse
 import * as semver from "semver"
 import { URL } from "url"
 import { AppUpdater } from "../AppUpdater"
-import { ResolvedUpdateFileInfo } from "../main"
+import { ResolvedUpdateFileInfo } from "../exports"
 import { getChannelFilename, newBaseUrl, newUrlFromBase } from "../util"
 import { parseUpdateInfo, Provider, ProviderRuntimeOptions, resolveFiles } from "./Provider"
